requirements.txt
async-timeout==4.0.3
grpcio==1.60.0
grpcio-tools==1.60.0
pick==2.2.0
protobuf==4.25.2
redis==5.0.1
windows-curses==2.3.2